+++ src/main/java/com/parkify/vehitrack/service/VehicleService.java
@@ -1,0 +1,8 @@
+package com.parkify.vehitrack.service;
+
+import com.parkify.vehitrack.entity.Vehicle;
+
+public interface VehicleService {
+
+    Vehicle createVehicle(Vehicle vehicle);
+}